+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # 版本更新记录
 
+## 2019-04-21 01:02:17
+
+> version:0.0.10
+
+1. `gaara build` 构建 prod 资源， `gaara build -N` 构建资源，但不上传 CDN
+2. `gaara dev [inputPath]` 支持指定构建某个目录的页面,不传则默认构建全部
+
+## 2019-04-20 17:11:24
+
+> version：0.0.9
+
+1. 新增 plugin-transform-runtime
+
 ## 2019-04-13 21:18:40
+
+> version：0.0.8
 
 1. 升级 webpack 版本， webpack4.29.6 在 React.Lazy 懒加载时报错 因此升级 4.29.6 =》 4.30.0